feat: use BatchRequest in DryRun
